Assignment-2/StockWarehouse/requirements.txt
asgiref==3.8.1
black==25.1.0
Brotli==1.1.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cssselect2==0.8.0
Django==5.1.7
djangorestframework==3.16.0
fonttools==4.56.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
weasyprint==65.0
webencodings==0.5.1
zopfli==0.2.3.post1
